pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "moresca"
dynamic = ["version"]
readme = "README.md"
description = "Reproducible boilerplate-free workflow management for Scanpy-based scRNA-seq analysis."
license = { text = "AGPL-3.0 license" }
requires-python = ">=3.11"

authors = [
    { name = "Matthias Bruhns", email = "matthias.bruhns@posteo.de" },
    { name = "Jan Thomas Schleicher", email = "jan.schleicher@uni-tuebingen.de" },
]

keywords = ["workflow management", "scRNA-seq", "reproducibility"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "scanpy",
    "doubletdetection",
    "harmonypy",
    "openpyxl",
    "anticor-features",
    "numpy",
    "gin-config",
    "loompy",
    "leidenalg",
    "igraph",
    "scikit-misc",
    "pynndescent",
    "triku@git+https://github.com/mbruhns/triku",
    "hotspotsc@git+https://github.com/YosefLab/Hotspot.git@workflow",
]

[project.urls]
repository = "https://github.com/claassenlab/MORESCA"

[project.optional-dependencies]
dev = ["ruff", "pytest", "pre-commit"]

[tool.setuptools_scm]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Assume Python 3.12.
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
extend-select = ["I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = true

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

#SOURCE: https://stackoverflow.com/a/76739181
[tool.setuptools.packages.find]
where = ["."] # list of folders that contain the packages (["."] by default)
include = [
    "MORESCA",
] # package names should match these glob patterns (["*"] by default)
exclude = [] # exclude packages matching these glob patterns (empty by default)
namespaces = false # to disable scanning PEP 420 namespaces (true by default)

[project.scripts]
moresca = "MORESCA.main:main"
